Fix faulty logic statement
In previous version, we use logical AND to distinguish if motor need to
be stopped at SUCCESSIVE mode and has `x_emer_flag` set, that will
cause motor keep operating if only one byte (DACxH or DACxL) is not
zero. Therefore, we fix it with logical OR to make the judgement work
properly.